requirements.txt
appnope==0.1.0
asgiref==3.3.1
backcall==0.2.0
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
decorator==4.4.2
Django==3.1.3
django-cors-headers==3.5.0
django-extensions==3.0.9
djangorestframework==3.12.2
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.2
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
traitlets==5.0.5
urllib3==1.26.2
wcwidth==0.2.5
